macplnk builds on Xcode 3.2.6 for x86_64. i386 would build expect the portaudio lib included is only 64-bit. Must fix that. Probably put all external libs (Ogg Vorbis)  in /ext then users could more easily avoid having to checkout those files

+++ src/plank/fft/plank_FFT.c
@@ -138,7 +138,7 @@ PlankResult pl_FFTF_InitWithLength (PlankFFTFRef p, const PlankL length)
     p->ifftScale = 0.5f;
 #else   
     p->peer = pl_FFTRealF_CreateAndInitWithLength (p->length);
-    p->fftScale = 2.f / p->length;
+    p->fftScale = 2.f / (int)p->length;
     p->ifftScale = 0.5f;
 #endif
